feat(score): 实现成绩管理功能和数据库连接池优化
- 添加 IllegalScoreException 异常类用于成绩验证
- 在 JdbcUtils 中实现 ThreadLocal 连接池管理
- 添加 closeConnection 方法用于连接资源释放
- 将 Score 实体的成绩字段从 BigDecimal 改为 Double
- 修改 ScoreDao insert 方法返回 Score 对象而非影响行数
- 在 ScoreServiceImpl 中添加成绩范围验证逻辑
- 修改 findScoresByStudentId 为 findScoresByStudentNo 方法
- 更新 StudentDao 返回 Student 对象并添加 countByCondition 方法
- 实现分页查询的学生数据传输对象 StudentPageDTO
- 在 StudentService 中使用 Optional 包装学生查询结果
- 实现学生和成绩的事务性删除操作
- 完成分页查询和批量导入功能实现

+++ src/main/java/week3/entity/Score.java
@@ -1,19 +1,18 @@
 package week3.entity;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Score {
     private Integer id;
     private Integer studentId;
     private String subject;
-    private BigDecimal score;
+    private Double score;
     private LocalDate examTime;
 
     public Score() {
     }
 
-    public Score(Integer studentId, String subject, BigDecimal score, LocalDate examTime) {
+    public Score(Integer studentId, String subject, Double score, LocalDate examTime) {
         this.studentId = studentId;
         this.subject = subject;
         this.score = score;
@@ -44,11 +43,11 @@ public class Score {
         this.subject = subject;
     }
 
-    public BigDecimal getScore() {
+    public Double getScore() {
         return score;
     }
 
-    public void setScore(BigDecimal score) {
+    public void setScore(Double score) {
         this.score = score;
     }
